+++ spec/spec_helper.rb
@@ -43,9 +43,11 @@ end
 
 require "wild"
 require_relative "support/wild_hooks/hook_fixtures"
+require_relative "support/wild_skillops/fixtures"
 
 RSpec.configure do |config|
   config.include Wild::Hooks::TestSupport::HookFixtures
+  config.include Wild::Skillops::TestSupport::Fixtures
   config.expect_with :rspec do |expectations|
     expectations.include_chain_clauses_in_custom_matcher_descriptions = true
   end